requirements.txt
pyyaml==6.0.2
pandas==2.3.1
numpy==2.2.6
scipy==1.15.2
matplotlib==3.10.3
seaborn==0.13.2
shap==0.48.0
scikit-learn==1.5.1
xgboost==3.0.5
catboost==1.2.8